[adhoc] Move aws IT tests to testing package according to best practices
